Worked on NrOfDays settings and made a file for theNrOfDays screen
