+++ lib/screens/weekplan_screen.dart
@@ -19,7 +19,7 @@ import 'package:weekplanner/models/enums/weekplan_mode.dart';
 import 'package:weekplanner/models/user_week_model.dart';
 import 'package:weekplanner/routes.dart';
 import 'package:weekplanner/screens/pictogram_search_screen.dart';
-import 'package:weekplanner/screens/settings_screen.dart';
+import 'package:weekplanner/screens/settings_screens/settings_screen.dart';
 import 'package:weekplanner/screens/show_activity_screen.dart';
 import 'package:weekplanner/widgets/bottom_app_bar_button_widget.dart';
 import 'package:weekplanner/widgets/giraf_app_bar_widget.dart';
@@ -282,14 +282,10 @@ class WeekplanScreen extends StatelessWidget {
                 builder: (BuildContext context,
                     AsyncSnapshot<SettingsModel> settingsSnapshot) {
 
-
                   if (settingsSnapshot.hasData) {
+                    final SettingsModel _settingsModel = settingsSnapshot.data;
+                    final int _daysToDisplay = _settingsModel.nrOfDaysToDisplay;
 
-                    SettingsModel settingsModel = settingsSnapshot.data;
-
-                    int _daysToDisplay = settingsModel.nrOfDaysToDisplay;
-
-                    print('DAAAAAAGE: ' + _daysToDisplay.toString());
                     // If the option of showing 1 day is chosen the
                     // _weekdayCounter
                     // must start from today's date
@@ -309,7 +305,6 @@ class WeekplanScreen extends StatelessWidget {
                       }
                     }
                   }
-
                   return Row(children: weekDays);
                 },
               );
